requirements.txt
python-dotenv
beautifulsoup4
spotipy
requests